edit bug fix

+++ src/controllers/videoController.js
@@ -1,14 +1,13 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async(req, res) => {
     const videos = await Video.find({}).sort({createdAt:"desc"});
-    console.log(videos);
     return res.render("home", {pageTitle:"Home", videos});
 };
 export const watch = async (req, res) => {
     const {id} = req.params;
     const video = await Video.findById(id).populate("owner");   // populate를 이용해서 owner에 있는 User의 데이터들을 가져온다.
-    console.log(video);
     if(video === null){
         return res.render("404", {pageTitle: "Video not found."});
     }
@@ -17,9 +16,13 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(video === null){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(video.owner != _id){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Edit: ${video.title} `, video});
 };
@@ -27,9 +30,13 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id:id});
-    if(video === null){
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(video.owner != _id){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,description, hashtags: Video.formatHashtags(hashtags),
@@ -46,13 +53,16 @@ export const postUpload = async (req, res) => {
     const {path:fileUrl} = req.file;
     const {title, description, hashtags} = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title: title,
             description,
             fileUrl,
             owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     }catch(error){
         return res.status(400).render("upload", {
@@ -64,7 +74,19 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) =>{
     const {id} = req.params;
+    const {user: {_id}, user} = req.session;
+    const video = await Video.findById(id).populate("owner");
+    const owner = video.owner;
+    if(video === null){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(owner._id != _id){
+        return res.status(403).redirect("/");
+    }
+    const {videos} = owner;
+    await videos.splice(videos.indexOf(id),1);
     await Video.findByIdAndDelete(id);
+    owner.save();
     return res.redirect("/");
 };
